Split text into script runs
Before this change, fontdiff asked ICU for the most likely script
given the BCP47 language tag of the text run. After this change,
fontdiff asks HarfBuzz for the actual script code for the rendered
codepoints.

The handling of common and inherited scripts is a bit simplistic: we
just take the script of the predecessing run, while eg. Raqm would try
to match opening/closing bidi pairs so they'd get the same script
code. But the new code is already much better than what we had before.

Resolves https://github.com/googlei18n/fontdiff/issues/7.