Score variation fonts by distance to supported range
Before this change, every font was assumed to have a single
width and weight, and we computed the distance to that value
when finding a good font to render a CSS style.  After this
change, fonts can have a range of supported widths and weights,
and we compute the distance to the range for scoring.